# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Rønningen Event Builder API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Error {
  error: string;
}

export type ServiceCategory = typeof ServiceCategory[keyof typeof ServiceCategory];


export const ServiceCategory = {
  venue: 'venue',
  food: 'food',
  transport: 'transport',
  activity: 'activity',
  service: 'service',
} as const;

export type ServicePricingType = typeof ServicePricingType[keyof typeof ServicePricingType];


export const ServicePricingType = {
  fixed: 'fixed',
  per_person: 'per_person',
  per_hour: 'per_hour',
  per_unit: 'per_unit',
  tiered: 'tiered',
  manual: 'manual',
} as const;

export interface Service {
  id: number;
  slug: string;
  name: string;
  category: ServiceCategory;
  pricingType: ServicePricingType;
  priceNok: number;
  costNok: number;
  minPriceNok: number;
  description: string;
  included?: string[];
  active: boolean;
}

export interface Supplier {
  id: number;
  name: string;
  serviceCount: number;
  preferred: boolean;
  active: boolean;
  /** @nullable */
  nextReview?: string | null;
}

export type QuoteMarketStatus = typeof QuoteMarketStatus[keyof typeof QuoteMarketStatus];


export const QuoteMarketStatus = {
  competitive: 'competitive',
  above_target: 'above_target',
  manual_review: 'manual_review',
} as const;

export type QuoteStatus = typeof QuoteStatus[keyof typeof QuoteStatus];


export const QuoteStatus = {
  draft: 'draft',
  approval: 'approval',
  sent: 'sent',
  accepted: 'accepted',
  declined: 'declined',
  changes_requested: 'changes_requested',
} as const;

export interface QuoteItem {
  id: number;
  name: string;
  category: string;
  quantity: number;
  unit?: string;
  unitPriceNok: number;
  totalNok: number;
  included: boolean;
}

export interface Quote {
  id: number;
  quoteNumber: string;
  company: string;
  /** @nullable */
  contactName?: string | null;
  eventDate: string;
  startTime?: string;
  endTime?: string;
  guests: number;
  eventType: string;
  totalNok: number;
  pricePerPersonNok: number;
  costNok: number;
  marginPercent: number;
  marketStatus: QuoteMarketStatus;
  status: QuoteStatus;
  items: QuoteItem[];
  expiryDate: string;
  createdAt?: string;
}

export interface DashboardSummary {
  openQuotes: number;
  pipelineNok: number;
  averageMargin: number;
  approvalCount: number;
  recentQuotes: Quote[];
}

export interface QuoteInput {
  /** @minLength 1 */
  company: string;
  contactName?: string;
  eventDate: string;
  startTime: string;
  endTime: string;
  /**
     * @minimum 1
     * @maximum 500
     */
  guests: number;
  eventType: string;
  /** @minItems 1 */
  serviceIds: number[];
}

export type QuoteStatusInputStatus = typeof QuoteStatusInputStatus[keyof typeof QuoteStatusInputStatus];


export const QuoteStatusInputStatus = {
  draft: 'draft',
  approval: 'approval',
  sent: 'sent',
  accepted: 'accepted',
  declined: 'declined',
  changes_requested: 'changes_requested',
} as const;

export interface QuoteStatusInput {
  status: QuoteStatusInputStatus;
}