# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Rønningen Event Builder API
 * OpenAPI spec version: 0.1.0
 */
import type { Quote } from './quote';

export interface DashboardSummary {
  openQuotes: number;
  pipelineNok: number;
  averageMargin: number;
  approvalCount: number;
  recentQuotes: Quote[];
}