# lib/api-zod/src/generated/types/quote.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Rønningen Event Builder API
 * OpenAPI spec version: 0.1.0
 */
import type { QuoteItem } from './quoteItem';
import type { QuoteMarketStatus } from './quoteMarketStatus';
import type { QuoteStatus } from './quoteStatus';

export interface Quote {
  id: number;
  quoteNumber: string;
  company: string;
  /** @nullable */
  contactName?: string | null;
  eventDate: Date;
  startTime?: string;
  endTime?: string;
  guests: number;
  eventType: string;
  totalNok: number;
  pricePerPersonNok: number;
  costNok: number;
  marginPercent: number;
  marketStatus: QuoteMarketStatus;
  status: QuoteStatus;
  items: QuoteItem[];
  expiryDate: Date;
  createdAt?: Date;
}