# lib/api-zod/src/generated/types/quoteInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Rønningen Event Builder API
 * OpenAPI spec version: 0.1.0
 */

export interface QuoteInput {
  /** @minLength 1 */
  company: string;
  contactName?: string;
  eventDate: Date;
  startTime: string;
  endTime: string;
  /**
     * @minimum 1
     * @maximum 500
     */
  guests: number;
  eventType: string;
  /** @minItems 1 */
  serviceIds: number[];
}