# lib/api-zod/src/generated/types/quoteItem.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Rønningen Event Builder API
 * OpenAPI spec version: 0.1.0
 */

export interface QuoteItem {
  id: number;
  name: string;
  category: string;
  quantity: number;
  unit?: string;
  unitPriceNok: number;
  totalNok: number;
  included: boolean;
}