# lib/api-zod/src/generated/types/quoteStatusInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Rønningen Event Builder API
 * OpenAPI spec version: 0.1.0
 */
import type { QuoteStatusInputStatus } from './quoteStatusInputStatus';

export interface QuoteStatusInput {
  status: QuoteStatusInputStatus;
}