# lib/api-zod/src/generated/types/service.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Rønningen Event Builder API
 * OpenAPI spec version: 0.1.0
 */
import type { ServiceCategory } from './serviceCategory';
import type { ServicePricingType } from './servicePricingType';

export interface Service {
  id: number;
  slug: string;
  name: string;
  category: ServiceCategory;
  pricingType: ServicePricingType;
  priceNok: number;
  costNok: number;
  minPriceNok: number;
  description: string;
  included?: string[];
  active: boolean;
}