# lib/api-zod/src/generated/types/supplier.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Rønningen Event Builder API
 * OpenAPI spec version: 0.1.0
 */

export interface Supplier {
  id: number;
  name: string;
  serviceCount: number;
  preferred: boolean;
  active: boolean;
  /** @nullable */
  nextReview?: Date | null;
}